requirements.txt
gdown
google-cloud-storage
python-dotenv
google-cloud-firestore
google-auth
from-root==1.3.0
pandas
numpy
python-box==7.3.2
PyYAML==6.0.2
-e .
